Fix for issue #115
Package slurm-torque-ohpc in the OpenHPC_1.0 repo cannot be installed
due to an unsatisfied dependency on slurm-perlapi. Note that the
OpenHPC_1.0 repo does contain a package "slurm-perlapi-ohpc", which
satisfies the dependencies on perl(Slurm) and perl(Slurmdb) according to
"yum deplist"; but the package appears to depend explicitly on
slurm-perlapi without the ohpc name, which is not available.

This change adds the PROJ_DELIM tag to the slurm-perlapi dependency of
slurm-torque-hpc, as specified in other dependencies.