Fix openblas aarch64 workaround
Apparently architecture detection isn't working right for
aarch64, but that also includes cpu core counts.  This maxes
out core count to 256 for now -- so that should keep us safe
in the short term, but we probably want to look into fixing
openblas arch detection upstream.  It is using /proc/cpuinfo
which is deprecated.  We should look to move it to sysfs if
later versions haven't already done this.

Signed-off-by: Eric Van Hensbergen <ericvh@gmail.com>