Replace '%define' with '%global' in %{!?PROJ_DELIM: %define PROJ_DELIM ...}
Do to strange rpm macro scoping rules a '%define' in a %{..: } construct
is only guaranteed to be set until the closing bracket:
https://www.redhat.com/archives/fedora-packaging/2009-May/msg00112.html
Therefore, replace it by '%global' - which is set until redefined.
Since %PROJ_DELIM is assigned a fixed value there is no difference between
'%define' and '%global' in this case.

Signed-off-by: Egbert Eich <eich@suse.com>