impi: Adjust version parsing to deal with spaces
This makes the version parsing replace " Update " with "." to prevent
spaces in generated module file names, which breaks other module
infrastructure.  This is specifically to address a problem with the impi
shipped with Parallel Studio Cluster Edition 2017 Update 1, in which
the version retrieved for automatic module file creation  was
"2017 Update 1". This now becomes "2017.1" thus resulting in a more
sensible module file name that doesn't break the remaining module
infrastructure..