macros: add repeated SPEC sections to the macros file
Instead of repeating the compiler selection code in each SPEC file this
commit moves the code (which is always almost the same) to the common
OHPC_macros file. This way it is easier to adapt the code as it only needs
to be changed in a single location. This is also in preparation of the
following patches to add support for gcc-6 and gcc-7.

Signed-off-by: Adrian Reber <areber@redhat.com>