ohpc-filesystem: new base packages for building and installing
Most of the openHPC spec files are owning the top-level directories in
/opt/ohpc. This means that during uninstall all packages are trying to
remove the top-level directories. This does not happen if other packages
are owning the directories. Nevertheless it is more common for RPM
packages to actually only own the directories it also uses. Therefore
no packages should own those top-level directories in the future.

In addition the ohpc-filesystem has a sub-package called ohpc-buildroot
which contains basic files and requires to build the openHPC RPMs.

This is the requirement on environment modules (lmod-ohpc) which seems
to be an implicit Requires: of OBS. Putting it into a RPM which is
installed in the build-root make the build more transparent and
independent of OBS.

Additionally the ohpc-buildroot package contains the files
OHPC_setup_compiler and OHPC_setup_mpi which used to be part of many
openHPC packages. Putting those files in a package which is always in
the build-root removes the requirement to provide those files in every
RPM package separately.

Signed-off-by: Adrian Reber <areber@redhat.com>