use the previously defined macros
This commit now adapts all openHPC spec files to use the new macros.

%ohpc_compiler and %ohpc_setup_compiler

This reduces the number of lines in the spec files significantly:
41 files changed, 607 insertions(+), 1702 deletions(-)

In addition unnecessary parts of the spec files have been removed.
ldconfig never needs to be executed as the libraries are not in paths
ldconfig looks into. '%clean' section are also not required in RPM for a
long time. BuildRoot is also always set by RPM automatically.

Signed-off-by: Adrian Reber <areber@redhat.com>