fix building with multiple Open MPIs
There is a small number of spec files which have different build options
depending on the used mpi_familiy. If neither mpich, mvapich2 or openmpi
is selected those %if-%endif blocks fail.

This commit adds additional quotes to the %if-%endif (which is
recommended) and it adds a new global variable called 'openmpi' which
returns just openmpi without v1.10, v2.0, v2.1 or v3.0.

With this patch I was able to build all packages for all versions of
Open MPI.

Signed-off-by: Adrian Reber <areber@redhat.com>