fix cube and tau build errors
Building the OpenHPC packages in COPR (https://copr.fedorainfracloud.org/)
results in errors in the tau and cube package. The reason seems to be
related to moving the installation trees from /tmp to the actual
build-root. Reproducing this error locally was not possible, but instead
of moving the tree to '..../%{version}/../%{version}' the spec file now
directly moves the result to '..../'. Not totally understood but this
patch helps. At the same time this patch removes unnecessary RPM tags,
which are part of the common macros file.

Signed-off-by: Adrian Reber <areber@redhat.com>