random cleanups
While fixing builds outside of OBS a few trailing whitespaces were
cleaned up and RPM tags which are already defined in the global macros
file. This also replaces a few %define with %global as RPM complains if
the %define is used in a %if as it is not visible outside of that block.

Signed-off-by: Adrian Reber <areber@redhat.com>