travis: configure travis
Our initial travis test is a small python script which only looks at the
SPEC files of the pull request and checks if they include any lines we
do not need. Currently this includes:

 * '^Source999.*$' (That is reserved for OHPC_macros)
 * '^Source.*OHPC_macros$' (That is not necessary anymore)
 * '^%changelog.*' (We removed all changelogs)
 * '^DocDir.*' (Set by OHPC_macros)
 * '^BuildRoot.*' (Not necessary anymore)
 * '^%defattr\(-,root,root\).*' (Setting the default value is not necessary)

With this initial travis setup running we can easily include more and
more checks in the future.

Signed-off-by: Adrian Reber <areber@redhat.com>