Don't use biosdevname for predictable network interface names
There are two ways of doing the predictable network interface names
thing. The original biosdevname approach by Dell, and the newer scheme
by systemd/udev,
https://www.freedesktop.org/wiki/Software/systemd/PredictableNetworkInterfaceNames/
.

Most distros have dropped biosdevname (including RHEL upstream,
Fedora, see https://bugzilla.redhat.com/show_bug.cgi?id=965718 for
issues with biosdevname and why it was dropped), and are nowadays
using the udev naming convention.

RHEL 7, however, is a bit of a weird hybrid, in that in a default
installation the udev naming is used, except if it's a Dell system, in
which case biosdevname is used. This default logic can be overridden
with the biosdevname=0/1 kernel command line arguments. See
https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/7/html/networking_guide/sec-consistent_network_device_naming_using_biosdevname

For Warewulf, lets follow upstream and thus not force the use of
biosdevname, and instead use the udev scheme as specified by
net.ifnames=1 without the additional biosdevname=1 argument.

Signed-off-by: Janne Blomqvist <janne.blomqvist@aalto.fi>