slurm: Using perl can break SRPM creation
If running in mock which is used by Fedora's build system the
SRPM is rebuild using a minimal build root which does not include
perl and that breaks the slrum SRPM creation.

As the perl dependency in the SRPM creation is only used to
detect if mariadb-devel or mysql-devel is available, the perl
dependency has been removed as OpenHPC only runs on systems
with mariadb-devel.

Signed-off-by: Adrian Reber <areber@redhat.com>