Add information about OpenHPC based containers
(This is still using my private quay.io registry URLs, which can be
replaced later with the real location)

During SC19 there were a couple of questions about OpenHPC based
containers. Although it is easy to create an OpenHPC based container
(see Dockerfiles) it is even more convenient if it already exists.

Details about the containers can be found in the README.md file.

 * make # to build all container
 * make push # build and push all container to a registry
 * make check # run some simple checks using the containers

Signed-off-by: Adrian Reber <areber@redhat.com>