Avoid killing user sessions on slurm.epilog
This PR fixes two issues and a cosmetic issue on slur.epilog.clean
script:

1.
The cosmetic one is trivial, tab vs spaces.

2.
The script should not run for UIDs lower than 1000 instead of 100.
Every UID lower than 1000 is considered a system account, this PR
fixes the comparison on the guard clause.

3.
The last one is a special case when the HPC environment uses the
headnode also as a computer node, or on HPC systems where nodes
aren't available and a single big server takes care of everything.

To avoid killing legitimate user sessions, like graphical login or
ssh logini, we check if `slurmctld` is running on the same machine
as the epilogue script. If it's not running the just kill
everything as the old behavior. If `slurmctld` is running just exit
without killing the remaining processes.

Signed-off-by: Vinícius Ferrão <vinicius@ferrao.net.br>
Signed-off-by: Georgia Stuart <georgia.stuart@gmail.com>