Increase reliability of chronyd for compute nodes
This commit will enhance the reliability of time synchronization on compute
nodes if there's a preexisting clock difference.

By enabling the compute nodes to "iburst" on the server, it will synchronize
it's clock with headnode imediatelly after chrony (as a client) started.

To further enhance it's reliability we also mark the headnode chronyd server
as a local stratum. In an event of unproper time fetch from upstream time
servers the headnode will be the default timekeeper for it's compute nodes.

This changes modifies the behavior of software like munge, that refuses to
work if there's a significant clock skew between the server and the client.
Avoiding the cluster to be offline due to clock synchronization issues.

It's also a best practice recommended in the chronyd FAQ as a minimal
configuration:
https://chrony.tuxfamily.org/faq.html#_what_is_the_minimum_recommended_configuration_for_an_ntp_client

Signed-off-by: Vinícius Ferrão <vinicius@ferrao.net.br>