Fix markdown complaints
My editor complains about a couple of things not being standard or
unusual in README.md and CONTRIBUTING.md.

This does not change any content, it just changes markdown usage. The
final layout should almost be the same.

With this change 8 from 9 complaints in README.md are resolved and 2
complaints in CONTRIBUTING.md are resolved.

Signed-off-by: Adrian Reber <areber@redhat.com>