Update warewulf-ipmi to 3.10.0
Add a patch for openEuler because its `ldd` returns paths starting with
`/usr/lib**` while all other distros (e.g. Rocky and Ubuntu) return
`/lib**`, i.e. without the leading `/usr`. Those are symlinks anyway.

Set wwsrvdir to "/var/lib" to be able to pass the build.
The problem is that if we use a custom value then it is ignored. The
configure line looks like:

```
[   22s] + ./configure --build=aarch64-openEuler-linux-gnu --host=aarch64-openEuler-linux-gnu --program-prefix= --disable-dependency-tracking --prefix=/usr --exec-prefix=/usr --bindir=/usr/bin --sbindir=/usr/sbin --sysconfdir=/etc --datadir=/usr/share --includedir=/usr/include --libdir=/usr/lib64 --libexecdir=/usr/libexec --localstatedir=/var --sharedstatedir=/var/lib --mandir=/usr/share/man --infodir=/usr/share/info --localstatedir=/var/lib --sharedstatedir=/var/lib --with-local-ipmitool=yes
```
Note that it contains --localstatedir and --sharedstatedir twice. The
second ones are from `%configure --localstatedir=%{wwsrvdir} --sharedstatedir=%{wwsrvdir} %{?CONF_FLAGS}`
but they seem to be always ignored.
Without this change the `%files initramfs-%{_arch}` at the common of the
spec fail to find the directories and the file

Signed-off-by: Martin Tzvetanov Grigorov <mgrigorov@apache.org>