Try to publish test run results in PR
With this commit tests on RHEL 9 (and clones) will generate junit
compatible output that needs to be uploaded by the test. Once the
files are uploaded and all tests are finished a new workflow
(test-analysis.yml) will use the uploaded files to generate an overview
in the PR about how many tests were running and successful.

Getting junit output on RHEL 9 (and clones) is a bit complicated as one
perl module does not exist in EPEL 9 which is needed. This change force
installs the corresponding RPM from EPEL 8.

Signed-off-by: Adrian Reber <areber@redhat.com>