tests/ci: simplify CI setup script
The CI setup script still contained a couple of if 8 or 9 code paths
which have been removed. This change also replaces of common used URLs
and paths with variables.

Once we have a release repository this script needs to be adapted once
more to use the release repository in combination with the latest
factory project on OBS.

Signed-off-by: Adrian Reber <areber@redhat.com>