compiler-families/intel-compilers-devel: require exact version
The current setup builds against the latest compiler available during
build time, but our packages are requiring a minimum version which might
not have all the symbols the compiler has we are using during build
time.

With this change we are requiring once specific version during build
time and during installation. This reduces the flexibility of our users
but ensures that that version of the packages are installed which were
used during build time.

Signed-off-by: Adrian Reber <areber@redhat.com>