Ignore rpmlint errors for packages in #1983
Building the newly added packages may fail because of rpmlint errors.
The most important ones are devel files in non-devel packages, which is
caused by OpenHPC not having devel packages, library-without-ldconfig-postin
and library-without-ldconfig-postun, which we do not want to run as
these packages are loaded via modules. Therefore, just ignore these
errors.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>